Close out body after reading if available
In more recent versions of Rack, if the body provides a `#close` method,
that is expected to be called after iteration over the contents of the
body. (See https://github.com/rack/rack/blob/ed9050876dfa2418613db05fb52015dc97cb78fb/SPEC#L250-L252
for the relevant section of the spec).